Reuse caching file snapshotter for build scripts
This commit introduces reuse of the caching file snapshotter for build scripts. A consequence
is that this caching file snapshotter is now always created, independently of the fact we're
using the daemon or not. The path to get a compiled script however still includes getting the
contents of a script, just in case the file is "empty".